Simplify getNodeIndex
This also happens to better work around the IE bug noted in the comment,
in that it still works in IE10PP2 while the old workaround broke.  Of
course, it might break in other ways if IE has more bizarre DOM bugs,
but hopefully not in the form of an infinite loop this time.