Commit masses of stuff
If I were expecting anyone to review this in detail, I'd probably break
it up into like ten commits.  The major thing is a new concept of
state/value override is added, partially fixing an important todo and
making commands like bold work for collapsed selections.  This will be
followed later by augmenting insertText to respect state/value for newly
inserted text.  Stuff I happened to do in the course of that, in the
order it appears in the diff:

* Don't treat hiliteColor specially for IE now that we support
  backColor.
* Record indeterm, state, and value in the execCommand() tests, and
  sanity-check them with some useful invariants that now mostly hold
  (after the overrides were added).
* Handle Firefox's insane treatment of backColor better.
* Make parseSimpleColor() vastly more robust.
* Link to the CSS3 extended color definitions.
* Add an XXX to foreColor, in a part that seems to be vague and/or not
  match implementations.
* Update color-related tests to use blue and brown for test colors
  instead of red and green, so it's easier to tell at a glance if the
  new indeterm/state/value indicators are complaining.
* Fix my usage of percentage rgb() and of hsl(), which were both syntax
  errors and therefore were not working.  rgb() needs either all
  arguments percentages or none of them, and hsl() needs the first
  argument not a percentage.
* Remove long-outdated note claiming backColor isn't specced.