Simplify "effectively contained" definition
If you examine the previous definition, you'll find that if the range is
collapsed, nothing can be effectively contained in it.  Thus I can move
four separate mentions of being collapsed to one at the beginning
instead.  The latter two mentions were totally redundant, since they
required that the range be collapsed, which was impossible under the
circumstances.

Tim Down brought this to my attention by pointing out misnested
parentheses in the implementation.  After thinking about it, I realized
that a) the misnesting only had the effect that the last range.collapsed
check did nothing, and b) this error would never actually affect the
result, so I could remove it from the definition.

Reported-By: Tim Down