Remove a line that's now basically redundant
Previously we needed the special case because we cared about
vertical-align and that only applies (in the way we care about) to
inline nodes, but now we don't care about vertical-align, so no need for
the extra check.  Authors who set sub/sup to display other than inline
will get what they deserve.