Revamp visibility and block-extend
I started out intending only to change the definition of invisibility to
exclude display: none and collapsed whitespace nodes, but the definition
of collapsed whitespace nodes turned out to be complicated and
indirectly led to me rewriting the block-extend algorithm.  The new
algorithms in the block-extend section are slightly shorter, but IMO a
lot more understandable.

Fixes: http://www.w3.org/Bugs/Public/show_bug.cgi?id=13631
Reported-By: Ryosuke Niwa
Report-URL: http://lists.whatwg.org/htdig.cgi/whatwg-whatwg.org/2011-August/032782.html