Fix a subtle bug in restoring overrides
This was causing the multitest ["foo[]bar", ["fontsize", "3"],
"subscript", "inserttext"] to fail.  When doing the inserttext, it first
produced "foo[a]bar", then "foo<sub>[a]</sub>bar".  But at this point,
the variable "node" was the text node "foo", not "a", since the node was
split.  This led to the fontsize command doing nothing, since the node
was already the correct size.